transaction success page, logging transactions to database

+++ src/main/java/com/cardealer/models/Transaction.java
@@ -42,7 +42,7 @@ public class Transaction {
     private double tax;
 
     @Column(name = "totalAmount")
-    private BigDecimal totalAmount;
+    private double totalAmount;
